+++ src/async/newest.rs
@@ -33,6 +33,7 @@ use semver::Version;
 /// ```
 ///
 /// [Crates.io]: https://crates.io/
+#[deprecated(since = "1", note = "Please use Versions struct")]
 pub async fn get_newest_version(
     crate_name: &str,
     current_crate_version: &str,