+++ src/blocking/max.rs
@@ -27,6 +27,7 @@ use semver::Version;
 /// ```
 ///
 /// [Crates.io]: https://crates.io/
+#[deprecated(since = "1", note = "Please use Versions struct")]
 pub fn get_max_version(
     crate_name: &str,
     current_crate_version: &str,
@@ -80,6 +81,7 @@ pub fn get_max_version(
 /// ```
 ///
 /// [Crates.io]: https://crates.io/
+#[deprecated(since = "1", note = "Please use Versions struct")]
 pub fn get_max_minor_version(
     crate_name: &str,
     version: &str,
@@ -133,6 +135,7 @@ pub fn get_max_minor_version(
 /// ```
 ///
 /// [Crates.io]: https://crates.io/
+#[deprecated(since = "1", note = "Please use Versions struct")]
 pub fn get_max_patch(
     crate_name: &str,
     version: &str,