+++ src/blocking/newest.rs
@@ -31,6 +31,7 @@ use semver::Version;
 /// ```
 ///
 /// [Crates.io]: https://crates.io/
+#[deprecated(since = "1", note = "Please use Versions struct")]
 pub fn get_newest_version(
     crate_name: &str,
     current_crate_version: &str,